// types/imgur.d.ts
// Generated by https://quicktype.io

export interface ImgurGallerySearchResponse {
  data: Gallery[];
  success: boolean;
  status: number;
}

export interface Gallery {
  id: string;
  title: string;
  description: null | string;
  datetime: number;
  cover?: string;
  cover_width?: number;
  cover_height?: number;
  account_url: string;
  account_id: number;
  privacy?: Privacy;
  layout?: Layout;
  views: number;
  link: string;
  ups: number;
  downs: number;
  points: number;
  score: number;
  is_album: boolean;
  vote: null;
  favorite: boolean;
  nsfw: boolean;
  section: string;
  comment_count: number;
  favorite_count: number;
  topic: null;
  topic_id: number | null;
  images_count?: number;
  in_gallery: boolean;
  is_ad: boolean;
  tags: Tag[];
  ad_type: number;
  ad_url: string;
  in_most_viral: boolean;
  include_album_ads?: boolean;
  images?: Image[];
  ad_config: AdConfig;
  type?: Type;
  animated?: boolean;
  width?: number;
  height?: number;
  size?: number;
  bandwidth?: number;
  has_sound?: boolean;
  edited?: number;
}

export interface AdConfig {
  safeFlags: SafeFlag[];
  highRiskFlags: any[];
  unsafeFlags: UnsafeFlag[];
  wallUnsafeFlags: string[];
  showsAds: boolean;
  showAdLevel: number;
  safe_flags: SafeFlag[];
  high_risk_flags: any[];
  unsafe_flags: UnsafeFlag[];
  wall_unsafe_flags: string[];
  show_ads: boolean;
  show_ad_level: number;
  nsfw_score: number;
}

export enum SafeFlag {
  Album = "album",
  InGallery = "in_gallery",
}

export enum UnsafeFlag {
  Mature = "mature",
  OnsfwModUnsafe = "onsfw_mod_unsafe",
  Under10 = "under_10",
}

export interface Image {
  id: string;
  title: null | string;
  description: null | string;
  datetime: number;
  type: Type;
  animated: boolean;
  width: number;
  height: number;
  size: number;
  views: number;
  bandwidth: number;
  vote: null;
  favorite: boolean;
  nsfw: null;
  section: null;
  account_url: null;
  account_id: null;
  is_ad: boolean;
  in_most_viral: boolean;
  has_sound: boolean;
  tags: any[];
  ad_type: number;
  ad_url: string;
  edited: string;
  in_gallery: boolean;
  link: string;
  comment_count: null;
  favorite_count: null;
  ups: null;
  downs: null;
  points: null;
  score: null;
}

export enum Type {
  ImageJPEG = "image/jpeg",
  ImagePNG = "image/png",
}

export enum Layout {
  Blog = "blog",
  Grid = "grid",
  Vertical = "vertical",
}

export enum Privacy {
  Hidden = "hidden",
  Public = "public",
}

// export enum Section {
//   Art = "Art",
//   Artcommissions = "artcommissions",
//   Empty = "",
//   HungryArtists = "HungryArtists",
//   Learntodraw = "learntodraw",
//   LoveLive = "LoveLive",
//   Pics = "pics",
// }

export interface Tag {
  name: string;
  display_name: string;
  followers: number;
  total_items: number;
  following: boolean;
  is_whitelisted: boolean;
  background_hash: string;
  thumbnail_hash: ThumbnailHash | null;
  accent: null | string;
  background_is_animated: boolean;
  thumbnail_is_animated: boolean;
  is_promoted: boolean;
  description: string;
  logo_hash: null;
  logo_destination_url: null;
  description_annotations: DescriptionAnnotations;
}

export interface DescriptionAnnotations {}

export enum ThumbnailHash {
  BFfAJEN = "BFfAJEN",
  The4ZY00KM = "4zY00kM",
}